refactor: Move navigation components to separate directory
