+++ src/features/navigation/TopAppBar.js
@@ -5,7 +5,7 @@ import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
 import IconButton from '@material-ui/core/IconButton';
 import MenuIcon from '@material-ui/icons/Menu';
-import { drawerWidth } from './constants';
+import { drawerWidth } from '../../constants';
 
 const useStyles = makeStyles(theme => ({
   root: {
